requirements.txt
aiohttp==3.5.4
APScheduler==3.6.0
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
discord.py==1.2.5
idna==2.8
idna-ssl==1.1.0
lxml==4.3.4
multidict==4.5.2
numpy==1.15.2
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.25.3
validators==0.13.0
websockets==6.0
yarl==1.3.0
